Extract Repo to own behaviour.
